// src/types/node.ts
export type NodeBody = {
  [x: string]: unknown;
};

export interface EdgeType {
  id: string;
  name: string;
  color: string;
}

export type NodeBodySubmit = {
  project_type_property_id: string;
  project_type_property_type: string;
  nodes_data: (string | number)[];
};

export type NodeDataSubmit = {
  project_id?: string;
  project_type_id: string;
  nodes?: NodeBodySubmit[];
  edges?: {
    id: string;
    project_edge_type_id: string;
    target_id: string;
    target_type_id: string;
    source_id: string;
    source_type_id: string;
  }[];
  nodeId?: string;
  name: string;
  default_image: string;
  destroyedEdgesIds?: string[];
};

export type ResponseLocationType = {
  address: string;
  location: {
    latitude: number;
    longitude: number;
  };
};

export type UploadedFileType = {
  name?: string;
  url: string;
};

export type NodeDataType = ResponseLocationType | string | number | boolean | UploadedFileType;

export type NodeDataTypes = NodeDataType[];

export type NodePropertiesValues = {
  id: string;
  node_id?: string;
  name: string[];
  default_image: UploadFile[];
  project_type_property_id?: string;
  project_type_property_type?: string;
  nodes_data?: NodeDataTypes;
  nodeTypeProperty: {
    name: string;
    id: string;
    default_image: boolean;
    default_property: boolean;
    multiple_type: boolean;
    enums_data: IEnumProperty[];
  };
  project_type_id: string;
  user_id: string;
  project_id: string;
  created_at: string;
  updated_at: string;
  createdEdges?: ICreateEdge[];
};
// ...
export interface ICreateEdge {
  id: string;
  label: string;
  source_id: string;
  target_id: string;
  project_edge_type_id: string;
}